\name{BSPADATA}
\alias{BSPADATA}
\docType{package}
\title{
Bayesian Spatial Econometric Models
}
\description{
This package includes 6 functions made to fit three Spatial Econometric Models proposed
  in Anselin(1988) in homoscedastic and heteroscedatic case. The fit is made through MCMC algorithms and working variables approach in the same fashion  as done in Cepeda(2001).
}
\details{
\tabular{ll}{
Package: \tab BSPADATA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-11-17\cr
License:\tab GPL (>=2)\cr
}

}
\author{
Jorge Sicacha-Parada <jasicachap@unal.edu.co> and Edilberto Cepeda-Cuervo <ecepedac@unal.edu.co>

Maintainer: Jorge Sicacha-Parada <jasicachap@unal.edu.co>
}
\references{
1. Cepeda C. E. (2001). Modelagem da variabilidade em modelos lineares generalizados. Unpublished Ph.D. tesis. Instituto de Matematicas.
Universidade Federal do Rio do Janeiro.

2.Cepeda, E. and Gamerman D. (2005). Bayesian Methodology for modeling parameters in the two-parameter exponential family. Estadistica 57, 93 105.

3.Cepeda C., E. and Gamerman D. (2001). Bayesian Modeling of Variance Heterogeneity in Normal Regression Models. Brazilian Journal of Probability and Statistics. 14, 207-221.

4.Luc Anselin, Spatial Econometrics: Methods and Models, Kluwer Academic, Boston, 1988.

5. D. Gamerman, Markov Chains Monte Carlo: Stochastic Simulation for bayesian Inference, Chapman and Hall, 1997.

6. James Le Sage and Kelley Pace, Introduction to Spatial Econometrics, Chapman & Hall/CRC, Boca Raton, 2009.
}
\keyword{Bayesian}
\keyword{Metropolis Hastings}
\keyword{Spatial Econometric Models}
\keyword{Heteroscedastic Error Term}
\keyword{Homoscedastic Error Term}


