% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero_general.R
\name{hetero_general_int}
\alias{hetero_general_int}
\title{Hello}
\usage{
hetero_general_int(
  y,
  X,
  Z,
  W1,
  W2 = NULL,
  nsim,
  burn,
  step,
  b_pri,
  B_pri,
  g_pri,
  G_pri,
  beta_0,
  gammas_0,
  rho_0,
  lambda_0,
  kernel = "normal",
  seed = 0,
  mateq = TRUE,
  impacts = TRUE
)
}
\description{
Hello
}
\keyword{internal}
