% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero_sar.R
\name{hetero_sar_int}
\alias{hetero_sar_int}
\title{Hello}
\usage{
hetero_sar_int(
  y,
  X,
  Z,
  W,
  nsim,
  burn,
  step,
  b_pri,
  B_pri,
  g_pri,
  G_pri,
  beta_0,
  gammas_0,
  rho_0,
  kernel = "normal",
  seed = 0,
  impacts = TRUE
)
}
\description{
Hello
}
\keyword{internal}
