% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero_sem.R
\name{hetero_sem_int}
\alias{hetero_sem_int}
\title{Hello}
\usage{
hetero_sem_int(
  y,
  X,
  Z,
  W,
  nsim,
  burn,
  step,
  b_pri,
  B_pri,
  g_pri,
  G_pri,
  beta_0,
  gammas_0,
  lambda_0,
  kernel = "normal",
  seed = 0
)
}
\description{
Hello
}
\keyword{internal}
