% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_general.R
\name{hom_general_int}
\alias{hom_general_int}
\title{Hello}
\usage{
hom_general_int(
  y,
  X,
  W1,
  W2 = NULL,
  nsim,
  burn,
  step,
  b_pri,
  B_pri,
  r_pri,
  lambda_pri,
  beta_0,
  sigma2_0,
  rho_0,
  lambda_0,
  kernel = "normal",
  seed = 0,
  mateq = TRUE,
  impacts = TRUE
)
}
\description{
Hello
}
\keyword{internal}
