% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_sar.R
\name{hom_sar_int}
\alias{hom_sar_int}
\title{Hello}
\usage{
hom_sar_int(
  y,
  X,
  W,
  nsim,
  burn,
  step,
  b_pri,
  B_pri,
  r_pri,
  lambda_pri,
  beta_0,
  sigma2_0,
  rho_0,
  kernel = "normal",
  seed = 0,
  impacts = TRUE
)
}
\description{
Hello
}
\keyword{internal}
