% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hom_sem.R
\name{hom_sem_int}
\alias{hom_sem_int}
\title{Hello}
\usage{
hom_sem_int(
  y,
  X,
  W,
  nsim,
  burn,
  step,
  b_pri,
  B_pri,
  r_pri,
  lambda_pri,
  beta_0,
  sigma2_0,
  lambda_0,
  kernel = "normal",
  seed = 0
)
}
\description{
Hello
}
\keyword{internal}
